fix: Move test to LoRA tests in dedicated file, use smaller model for entrypoint test Signed-off-by: Sanger Steel <sangersteel@gmail.com>
